Closer.

+++ src/test/scala/mimir/timing/vldb2017/ImputeTiming.scala
@@ -28,12 +28,13 @@ object ImputeTiming
   val runSamplerQueries = true
   val useMaterialized = false
   val useFastPathCache = true
-
+  val skipImputeTests = true
+  
   val timeout = 15.minute
 
   def beforeAll =
   {
-    if(fullReset){
+    if(fullReset && ! skipImputeTests){
       println("DELETING ALL MIMIR METADATA")
       update("DELETE FROM MIMIR_MODEL_OWNERS")
       update("DELETE FROM MIMIR_MODELS")
@@ -76,7 +77,7 @@ object ImputeTiming
       Seq("CUSTKEY")
     ))
   )
-  if(true){ "Skipping TPCH Inpute Test" >> ok } else {
+  if(skipImputeTests){ "Skipping TPCH Inpute Test" >> ok } else {
     "TPCH Impute" should {
 
       sequential
